requirements.txt
pretty-errors==1.2.25
psycopg2-binary==2.9.9
